+++ domain/src/main/java/pro/number/domain/usecase/GetParticipantsByGroupIdUseCase.kt
@@ -7,6 +7,6 @@ class GetParticipantsByGroupIdUseCase @Inject constructor(
     private val repository: ReceiptRepository
 ) {
 
-    suspend operator fun invoke(groupId: Int) = repository.getParticipantsByGroupId(groupId)
+    suspend operator fun invoke(groupId: Long) = repository.getParticipantsByGroupId(groupId)
 
 }
